pyproject.toml
[project]
name = "project1"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.14"
dependencies = ["langchain==1.2.10",
                "langgraph==1.0.10",
                "langchain_aws==1.3.1",
                "langchain_openai==1.1.10",
                "boto3==1.42.59",
                "faiss-cpu==1.13.2",
                "sentence-transformers==5.2.3"]
